Refactor out test Index classes
